+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Unused `getColorForKind` function that was replaced with go-pretty styling
+- Unused command implementations (create, read, update, delete, device) - only list command remains as it's the only implemented verb
 
 ## [1.0.0] - 2025-09-04